Added basic walk animation controller
- Added sampling of animations
- Added updating of bone matrices from rig model
- Re-orienting of the entity based on current velocity

+++ src/string_utils.h
@@ -1,10 +1,26 @@
-/*
- * MeshUp - A visualization tool for multi-body systems based on skeletal
- * animation and magic.
- *
- * Copyright (c) 2012 Martin Felis <martin.felis@iwr.uni-heidelberg.de>
- *
- * Licensed under the MIT license. See LICENSE for more details.
+/* 
+ * Puppeteer - A Motion Capture Mapping Tool
+ * Copyright (c) 2013-2015 Martin Felis <martin.felis@iwr.uni-heidelberg.de>.
+ * All rights reserved.
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining
+ * a copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, sublicense, and/or sell copies of the Software, and to
+ * permit persons to whom the Software is furnished to do so, subject to
+ * the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
+ * CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
+ * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+ * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE* 
  */
 
 #ifndef _STRING_UTILS_H
@@ -36,7 +52,7 @@ inline std::string strip_whitespaces (const std::string &line, std::string white
 
 inline std::string tolower (const std::string &line) {
 	std::string result (line);
-	for (int i = 0; i < line.size(); i++) 
+	for (size_t i = 0; i < line.size(); i++) 
 		result[i] = tolower(result[i]);
 
 	return result;
